[REFACTOR] 패키지 정리 및 불필요한 주석 삭제
