--- src/main/java/com/nextroom/oescape/config/security/JwtAuthenticationToken.java
@@ -1,27 +1,0 @@
-package com.nextroom.oescape.config.security;
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-
-public class JwtAuthenticationToken extends UsernamePasswordAuthenticationToken {
-
-    private String token;
-
-    public JwtAuthenticationToken(String token) {
-        super(null, null);
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    @Override
-    public Object getPrincipal() {
-        return null;
-    }
-
-    @Override
-    public Object getCredentials() {
-        return null;
-    }
-}
