+++ src/main/java/com/nextroom/oescape/controller/AuthController.java
@@ -6,8 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.nextroom.oescape.config.security.TokenDto;
-import com.nextroom.oescape.domain.Shop;
+import com.nextroom.oescape.dto.response.TokenResponseDto;
 import com.nextroom.oescape.dto.request.ShopRequestDto;
 import com.nextroom.oescape.dto.request.TokenRequestDto;
 import com.nextroom.oescape.dto.response.ShopResponseDto;
@@ -29,12 +28,12 @@ public class AuthController {
     }
 
     @PostMapping("/login")
-    public ResponseEntity<TokenDto> login(@RequestBody ShopRequestDto shopRequestDto) {
+    public ResponseEntity<TokenResponseDto> login(@RequestBody ShopRequestDto shopRequestDto) {
         return ResponseEntity.ok(authService.login(shopRequestDto));
     }
 
     @PostMapping("/reissue")
-    public ResponseEntity<TokenDto> reissue(@RequestBody TokenRequestDto tokenRequestDto) {
+    public ResponseEntity<TokenResponseDto> reissue(@RequestBody TokenRequestDto tokenRequestDto) {
         return ResponseEntity.ok(authService.reissue(tokenRequestDto));
     }
 }
