+++ src/main/java/com/nextroom/oescape/security/SecurityUtil.java
@@ -1,4 +1,4 @@
-package com.nextroom.oescape.config.security;
+package com.nextroom.oescape.security;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -10,8 +10,6 @@ public class SecurityUtil {
 
     private SecurityUtil() { }
 
-    // SecurityContext 에 유저 정보가 저장되는 시점
-    // Request 가 들어올 때 JwtFilter 의 doFilter 에서 저장
     public static Long getCurrentMemberId() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
