+++ src/main/java/com/nextroom/oescape/security/config/JwtSecurityConfig.java
@@ -1,18 +1,19 @@
-package com.nextroom.oescape.config.security;
+package com.nextroom.oescape.security.config;
 
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.DefaultSecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
+import com.nextroom.oescape.security.TokenProvider;
+import com.nextroom.oescape.security.jwt.JwtFilter;
+
 import lombok.RequiredArgsConstructor;
 
-// 직접 만든 TokenProvider 와 JwtFilter 를 SecurityConfig 에 적용할 때 사용
 @RequiredArgsConstructor
 public class JwtSecurityConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
     private final TokenProvider tokenProvider;
 
-    // TokenProvider 를 주입받아서 JwtFilter 를 통해 Security 로직에 필터를 등록
     @Override
     public void configure(HttpSecurity http) {
         JwtFilter customFilter = new JwtFilter(tokenProvider);
