+++ src/main/java/com/nextroom/oescape/security/jwt/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.nextroom.oescape.config.security;
+package com.nextroom.oescape.security.jwt;
 
 import java.io.IOException;
 
@@ -19,7 +19,6 @@ public class JwtAccessDeniedHandler implements AccessDeniedHandler {
         IOException,
         ServletException,
         IOException {
-        // 필요한 권한이 없이 접근하려 할때 403
         response.sendError(HttpServletResponse.SC_FORBIDDEN);
     }
 }
