+++ src/main/java/com/nextroom/oescape/security/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.nextroom.oescape.config.security;
+package com.nextroom.oescape.security.jwt;
 
 import java.io.IOException;
