+++ src/main/java/com/nextroom/oescape/service/CustomUserDetailsService.java
@@ -29,7 +29,6 @@ public class CustomUserDetailsService implements UserDetailsService {
             .orElseThrow(() -> new UsernameNotFoundException(username + " -> 데이터베이스에서 찾을 수 없습니다."));
     }
 
-    // DB 에 User 값이 존재한다면 UserDetails 객체로 만들어서 리턴
     private UserDetails createUserDetails(Shop shop) {
         GrantedAuthority grantedAuthority = new SimpleGrantedAuthority(shop.getAuthority().toString());
